PxWallpaper/requirements.txt
olefile==0.44
Pillow==4.0.0
requests==2.13.0
WMI==1.4.9

